Refectored using usecontext
