+++ src/store/auth-context.js
@@ -1,7 +1,43 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 
 const AuthContext = React.createContext({
   isLoggedIn: false,
+  onLogout: () => {},
+  onLogin: (email, password) => {},
 });
+
+export const AuthContextProvider = (props) => {
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+
+  const logoutHandler = () => {
+    localStorage.removeItem("isloggedIn");
+    setIsLoggedIn(false);
+  };
+
+  const loginHandler = () => {
+    localStorage.setItem("isloggedIn", "1");
+    setIsLoggedIn(true);
+  };
+
+  useEffect(() => {
+    const storedUserLoggedIn = localStorage.getItem("isLoggedIn");
+
+    if (storedUserLoggedIn === "1") {
+      setIsLoggedIn(true);
+    }
+  }, []);
+
+  return (
+    <AuthContext.Provider
+      value={{
+        isLoggedIn: isLoggedIn,
+        onLogout: logoutHandler,
+        onLogin: loginHandler,
+      }}
+    >
+      {props.children}
+    </AuthContext.Provider>
+  );
+};
 
 export default AuthContext;
